+++ imt_model/src/imtPSS_NoSurf.h
@@ -263,7 +263,7 @@ namespace Cantera
 
 
 
-    void printInterfacialMassTransfer_Phase(int iph, int pSrc, bool subTimeStep);
+    void printInterfacialMassTransfer_Phase(size_t iph, int pSrc, bool subTimeStep);
 
     //! Write out CSV tabular data on the integrations
     /*!
